portal/requirements.txt
Brotli==1.0.7
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Compress==1.5.0
Flask-JSON==0.3.4
Flask-Limiter==1.3.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
flask-talisman==0.7.0
Flask-WTF==0.14.3
haversine==2.3.0
idna==2.10
importlib-metadata==1.7.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==3.0.0
limits==1.5.1
MarkupSafe==2.0.0
mysqlclient==2.0.1
pycountry==20.7.3
python-dotenv==0.13.0
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.18
stripe==2.55.0
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
